WIP: feat dynamic roles / capabilities w/ only owner restriction (NOT-TESTED)
Introduced EnumerableSet.UintSet to easily manage role permissions as an enumerable set
Introduced RolePermissionsStruct: we couldn't call "delete rolePermissions[role]" at role deletion because it's no allowed to do so in nested-mappings
isLegitRole && notDefault modifiers to understand if a role is truly existant and if it's not default to prevent default ranks alteration / deletion
Introduced addRole && revokeRole (need to be heavily tested)

------

We might consider using a single structure for users, maybe a mapping? (address => UserStruct) that has tokenAuthorizations, crowdsaleManagement && exchangeManagement?